Move some interal.core test packages into core package.
- Related #376
- Move MicroProfileForJavaAssert, MicroProfileAssert and JavaUtils
  from the restricted packages into the already exported
  org.eclipse.lsp4mp.jdt.core .
