+++ microprofile.jdt/org.eclipse.lsp4mp.jdt.test/src/main/java/org/eclipse/lsp4mp/jdt/core/restclient/MicroProfileRegisterRestClientTest.java
@@ -13,13 +13,13 @@
 *******************************************************************************/
 package org.eclipse.lsp4mp.jdt.core.restclient;
 
-import static org.eclipse.lsp4mp.jdt.internal.core.MicroProfileAssert.assertHints;
-import static org.eclipse.lsp4mp.jdt.internal.core.MicroProfileAssert.assertHintsDuplicate;
-import static org.eclipse.lsp4mp.jdt.internal.core.MicroProfileAssert.assertProperties;
-import static org.eclipse.lsp4mp.jdt.internal.core.MicroProfileAssert.assertPropertiesDuplicate;
-import static org.eclipse.lsp4mp.jdt.internal.core.MicroProfileAssert.h;
-import static org.eclipse.lsp4mp.jdt.internal.core.MicroProfileAssert.p;
-import static org.eclipse.lsp4mp.jdt.internal.core.MicroProfileAssert.vh;
+import static org.eclipse.lsp4mp.jdt.core.MicroProfileAssert.assertHints;
+import static org.eclipse.lsp4mp.jdt.core.MicroProfileAssert.assertHintsDuplicate;
+import static org.eclipse.lsp4mp.jdt.core.MicroProfileAssert.assertProperties;
+import static org.eclipse.lsp4mp.jdt.core.MicroProfileAssert.assertPropertiesDuplicate;
+import static org.eclipse.lsp4mp.jdt.core.MicroProfileAssert.h;
+import static org.eclipse.lsp4mp.jdt.core.MicroProfileAssert.p;
+import static org.eclipse.lsp4mp.jdt.core.MicroProfileAssert.vh;
 
 import org.eclipse.lsp4mp.commons.MicroProfileProjectInfo;
 import org.eclipse.lsp4mp.commons.MicroProfilePropertiesScope;
